presence: Cast time_t
CID #417558, #417600, #417453, #417502, #417672, #417744
    #417756, #417782, #417812, #417578, #417444, #417539
	#417680, #417743, #417727, #417745, #417531, #417650
    #417711, #417776, #417822, #417826

+++ modules/presence/notify.c
@@ -389,7 +389,7 @@ str* get_wi_notify_body(subs_t* subs, subs_t* watcher_subs)
 	{
 		s= s->next;
 
-		if(s->expires< (int)time(NULL))
+		if(s->expires< (unsigned int)(unsigned long)time(NULL))
 		{
 			LM_DBG("expired record\n");
 			continue;
@@ -1597,10 +1597,10 @@ int get_subs_db(str* pres_uri, pres_ev_t* event, str* sender,
 			s.event= event;
 			s.local_cseq = row_vals[cseq_col].val.int_val;
 
-			if(row_vals[expires_col].val.int_val < (int)time(NULL))
+			if(row_vals[expires_col].val.int_val < (int)(unsigned long)time(NULL))
 				s.expires = 0;
 			else
-				s.expires = row_vals[expires_col].val.int_val -(int)time(NULL);
+				s.expires = row_vals[expires_col].val.int_val -(int)(unsigned long)time(NULL);
 			s.version = row_vals[version_col].val.int_val;
 
 			s_new= mem_copy_subs(&s, PKG_MEM_TYPE);
@@ -1660,7 +1660,7 @@ int update_in_list(subs_t* s, subs_t* s_array, int new_rec_no, int n)
 		strncmp(ls->from_tag.s, s->from_tag.s, s->from_tag.len)== 0 )
 		{
 			ls->local_cseq= s->local_cseq;
-			ls->expires= s->expires- (int)time(NULL);
+			ls->expires= s->expires- (unsigned int)(unsigned long)time(NULL);
 			ls->version= s->version;
 			ls->status= s->status;
 			return 1;
@@ -1692,7 +1692,7 @@ int presentity_has_subscribers(str* pres_uri, pres_ev_t* event)
 		s= s->next;
 
 		/* expired & active ? */
-		if ( (s->expires<(int)now) || (s->status!=ACTIVE_STATUS) ||
+		if ( (s->expires<(unsigned int)(unsigned long)now) || (s->status!=ACTIVE_STATUS) ||
 		(s->reason.len!=0) )
 			continue;
 
@@ -1807,7 +1807,7 @@ subs_t* get_subs_dialog(str* pres_uri, pres_ev_t* event, str* sender,
 
 			printf_subs(s);
 
-			if(s->expires< (int)time(NULL))
+			if(s->expires< (unsigned int)(unsigned long)time(NULL))
 			{
 				LM_DBG("expired subs\n");
 				continue;
@@ -1829,7 +1829,7 @@ subs_t* get_subs_dialog(str* pres_uri, pres_ev_t* event, str* sender,
 				lock_release(&subs_htable[hash_code].lock);
 				goto error;
 			}
-			s_new->expires-= (int)time(NULL);
+			s_new->expires-= (unsigned int)(unsigned long)time(NULL);
 			s_new->next= s_array;
 			s_array= s_new;
 			i++;
